fix test

+++ promql/engine_test.go
@@ -747,7 +747,6 @@ load 10s
   metricWith3SampleEvery10Seconds{a="2",b="2"} 1+1x100
   metricWith3SampleEvery10Seconds{a="3",b="2"} 1+1x100
 `)
-
 	require.NoError(t, err)
 	defer test.Close()
 
@@ -757,7 +756,7 @@ load 10s
 	cases := []struct {
 		Query               string
 		SkipMaxCheck        bool
-		TotalSamples        int
+		TotalSamples        int64
 		TotalSamplesPerStep stats.TotalSamplesPerStep
 		PeakSamples         int
 		Start               time.Time
